Switch to double for efficiency

+++ math/Real.h
@@ -1,6 +1,6 @@
 #ifndef SOLARSCAPE_REAL_H
 #define SOLARSCAPE_REAL_H
 
-using Real = long double;
+using Real = double;
 
 #endif
